+++ train.py
@@ -9,9 +9,10 @@ from tqdm import tqdm
 
 from nkb_classification.dataset import get_dataset
 from nkb_classification.losses import get_loss
-from nkb_classification.metrics import compute_metrics, log_confusion_matrices, log_metrics
+from nkb_classification.metrics import compute_metrics
 from nkb_classification.model import get_model
-from nkb_classification.utils import get_experiment, get_optimizer, get_scheduler, log_grads, log_images, read_py_config
+from nkb_classification.utils import get_experiment, get_optimizer, get_scheduler, read_py_config
+from nkb_classification.logging import TrainLogger
 
 
 class TrainPbar(tqdm):
@@ -26,121 +27,6 @@ class TrainPbar(tqdm):
             self.set_postfix_str(f"Loss: {loss['loss'].item():.4f}")
         else:
             self.set_postfix_str(f"Loss: {loss.item():.4f}")
-
-
-class TrainLogger:
-    __slots__ = (
-        "cfg",
-        "epoch_running_loss",
-        "epoch_confidences",
-        "epoch_predictions",
-        "epoch_ground_truth",
-        "target_names",
-        "label_names",
-        "experiment",
-        "task",
-        "class_to_idx",
-    )
-
-    def __init__(self, cfg, experiment, class_to_idx):
-
-        assert cfg.task in ("single", "multi")
-
-        self.cfg = cfg
-        self.experiment = experiment
-        self.task = cfg.task
-        self.class_to_idx = class_to_idx
-
-    def init_iter_logs(self):
-
-        if self.task == "single":
-            self.epoch_running_loss = []
-            self.epoch_confidences = []
-            self.epoch_predictions = []
-            self.epoch_ground_truth = []
-
-            self.target_names = None
-            # import ipdb; ipdb.set_trace()
-            self.label_names = list(self.class_to_idx.keys())
-
-        elif self.task == "multi":
-            self.epoch_running_loss = defaultdict(list)
-            self.epoch_confidences = defaultdict(list)
-            self.epoch_predictions = defaultdict(list)
-            self.epoch_ground_truth = defaultdict(list)
-
-            self.target_names = [*sorted(self.class_to_idx)]
-
-            self.label_names = {
-                target_name: [*self.class_to_idx[target_name].keys()] for target_name in self.target_names
-            }
-
-    def log_iter(self, pred, true, loss):
-        assert type(pred) == type(true)
-
-        # Multi-task case
-        if isinstance(pred, dict):
-            assert pred.keys() == true.keys()
-            for target_name in pred.keys():
-                self.epoch_ground_truth[target_name].extend(true[target_name].cpu().numpy().tolist())
-                self.epoch_confidences[target_name].extend(
-                    pred[target_name].softmax(dim=-1, dtype=torch.float32).detach().cpu().numpy().tolist()
-                )
-                self.epoch_predictions[target_name].extend(
-                    pred[target_name].argmax(dim=-1).detach().cpu().numpy().tolist()
-                )
-                self.epoch_running_loss[target_name].append(loss[target_name].item())
-            self.epoch_running_loss["loss"].append(loss["loss"].item())
-        else:
-            self.epoch_ground_truth.extend(true.cpu().numpy().tolist())
-            self.epoch_confidences.extend(pred.softmax(dim=-1, dtype=torch.float32).detach().cpu().numpy().tolist())
-            self.epoch_predictions.extend(pred.argmax(dim=-1).detach().cpu().numpy().tolist())
-            self.epoch_running_loss.append(loss.item())
-
-    def get_epoch_results(self):
-        return {
-            "running_loss": self.epoch_running_loss,
-            "confidences": self.epoch_confidences,
-            "predictions": self.epoch_predictions,
-            "ground_truth": self.epoch_ground_truth,
-        }
-
-    def log_epoch(self, epoch, train_results, val_results):
-        epoch_val_acc = None
-        if self.experiment is not None:  # log metrics
-            log_images(self.experiment, "Train", epoch, train_results["images"])
-
-            log_images(self.experiment, "Validation", epoch, val_results["images"])
-
-            log_metrics(
-                self.experiment,
-                self.target_names,
-                self.label_names,
-                epoch,
-                train_results["metrics"],
-                "Train",
-            )
-
-            log_metrics(
-                self.experiment,
-                self.target_names,
-                self.label_names,
-                epoch,
-                val_results["metrics"],
-                "Validation",
-            )
-
-            log_confusion_matrices(
-                self.experiment,
-                self.target_names,
-                self.label_names,
-                epoch,
-                val_results,
-                "Validation",
-            )
-
-            if self.cfg.log_gradients:
-                log_grads(self.experiment, epoch, train_results["metrics_grad_log"])
 
 
 def train_epoch(
